Tue, Jun 7, 2022, 11:49 PM +00:00

+++ filter-less/helpers.c
@@ -90,9 +90,9 @@ void reflect(int height, int width, RGBTRIPLE image[height][width])
 // Blur image
 void blur(int height, int width, RGBTRIPLE image[height][width])
 {
-    BYTE avgB;
-    BYTE avgG;
-    BYTE avgR;
+    int avgB;
+    int avgG;
+    int avgR;
     int surroundingCount;
 
     RGBTRIPLE imageCopy[height][width];
